Title: 삼 오 무, Time: 3027ms, Memory: 8MB, Status: Time Limit Exceed - CodeTree

+++ 231211/삼 오 무/three-five-moo.java
@@ -7,19 +7,19 @@ public class Main {
     static BufferedReader buffer = new BufferedReader(new InputStreamReader(System.in));
     static StringTokenizer tokens;
     
-    static int n; //점의 갯수 
+    static long n; //점의 갯수 
     
     
     static boolean isPossible(int k) {
     	
-    	int mooCount = k/3+k/5-k/15; 
+    	long mooCount = k/3+k/5-k/15; 
     	
     	return k-mooCount>=n;
     	
     }
     
     public static void main(String[] args) throws IOException{
-    	n = Integer.valueOf(buffer.readLine()); 
+    	n = Long.valueOf(buffer.readLine()); 
     	
     	
     	int left = 1;
